add drawPetal
